+++ website_sale_cart_expire/__manifest__.py
@@ -5,7 +5,7 @@
 {
     "name": "Website Sale Cart Expire",
     "summary": "Cancel carts without activity after a configurable time",
-    "version": "16.0.1.0.1",
+    "version": "17.0.1.0.0",
     "author": "Camptocamp, Odoo Community Association (OCA)",
     "maintainers": ["ivantodorovich"],
     "website": "https://github.com/OCA/e-commerce",
@@ -19,7 +19,7 @@
     ],
     "assets": {
         "web.assets_frontend": [
-            "website_sale_cart_expire/static/src/js/website_sale_cart_expire.js",
+            "website_sale_cart_expire/static/src/js/website_sale_cart_expire.esm.js",
         ],
     },
 }
